Probably make octree storage and memory manager work together again

+++ terra-offheap/src/main/java/com/ritualsoftheold/terra/offheap/io/OctreeLoader.java
@@ -21,5 +21,5 @@ public interface OctreeLoader {
      * @param index
      * @param addr
      */
-    void saveOctrees(byte index, long addr);
+    void saveOctrees(int index, long addr);
 }
